pyproject.toml
[project]
name = "openlogi-ai-faq"
version = "0.1.0"
description = "A project to crawl OpenLogi FAQ data and provide a Q&A interface using Gemini API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "google-generativeai>=0.8.4",
    "lxml>=5.3.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]
